minor adjustments to layout and spacing
